+++ RELEASENOTES.md
@@ -5,6 +5,7 @@ These are the release notes for the [wikibase-api](README.md).
 * Allows more versions of wikibase/data-model ~4.2|~5.0|~6.0|~7.0|~8.0
 * Allows version 0.7 of addwiki/mediawiki-datamodel
 * Uses EntityRetrievingTermLookup instead of TermApiLookup to implement TermLookup interface
+* Added newEntitySearcher method to WikibaseFactory
 
 ## Version 0.7 (2016-07-04)